requirements.txt
accelerate==0.16.0
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.0.10
filelock==3.9.0
flake8==6.0.0
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
mccabe==0.7.0
multidict==6.0.4
numpy==1.24.2
openai==0.27.0
packaging==23.0
Pillow==9.4.0
psutil==5.9.4
pycodestyle==2.10.0
pyflakes==3.0.1
python-dotenv==1.0.0
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
sqlparse==0.4.3
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.14
whitenoise==5.3.0
yarl==1.8.2
zipp==3.15.0
